pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "bev-vawa-lite"
version = "0.1.0"
description = "Lightweight BEV-native Vision-Action / World-Action navigation (PointGoal)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.4",
    "mujoco>=3.1",
    "numpy>=1.26",
    "pyyaml>=6.0",
    "tqdm>=4.65",
    "tyro>=0.8",
    "imageio>=2.30",
    "imageio-ffmpeg>=0.4",
    "matplotlib>=3.7",
]

[project.optional-dependencies]
dev = ["pytest>=7"]

[tool.setuptools.packages.find]
where = ["."]
include = ["bev_vawa*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
